Move ViewRef to views sdk.
